requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.0.4
dnspython==2.2.1
Flask==2.1.3
Flask-PyMongo==2.3.0
idna==3.4
itsdangerous==2.0.1
Jinja2==3.1.2
joblib==1.2.0
MarkupSafe==2.1.1
nltk==3.7
numpy==1.23.4
Pillow==9.3.0
pip==22.2.2
pymongo==4.3.2
python-dotenv==0.21.0
regex==2022.10.31
requests==2.28.1
setuptools==65.5.0
six==1.16.0
tflearn==0.5.0
torch==1.13.0
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.12
Werkzeug==2.0.3
wheel==0.37.1
